quirky state pattern implementation to represent Game's phase (planning, exection)
this was an experiment to try and make the Game's phase explicit without using an enum.
the motivation for this was I want to refactor UI where ui.switch_to_execute_mode() etc.
would be replaced by a unified ui.update_presentation(game)
which would read whatever data from game it needs to draw itself, including the phase.
however, there isn't really much reason for state pattern, although it's kind of neat
to have all the tick timer business out of game and into the execution phase,
because it's only used there.
i guess it would still be tricky for UI to determine the game phase without knowing what phases exist.
because it would need to ask
`if game.phase == game.execution_phase: ...`
or require some enum parameter to `switch(game.phase.id)` on.
So, all in all, it still seems more sensible to just provide ui.update(game_state_data)
with a data transfer object that contains an enum to tell the phase.