implemented explicit Commander "appendage objects" for UI and AIPlayer.
by appendage objects i mean "optional child" nodes of AI and UI, receiving signals from their parents and signalling directly to grandparent (Game).
The point being that the Commander concept is made explicit this way and if Commanders get rich functionality in the future, it's already extracted in this way.